==> Data_Preparation/03_eBird_taxonomy.R <==

library(tidyverse)
data.path <- "X:/morton_research/User/bi1om/Research/Wildlife_trade/Morton_et_al_eBird_Index/Analysis/"

# ID/MY sp. Taxonomy ---------------------------------------------------------------
ID_sp <- read.csv(paste0(data.path, "Data/eBird/Obs_species/ID_filt_all.csv")) %>% select(-X)
MY_sp <- read.csv(paste0(data.path, "Data/eBird/Obs_species/MY_filt_all.csv")) %>% select(-X)
all_sp <- rbind(ID_sp, MY_sp) %>% group_by(scientific_name, common_name) %>% tally()

# IUCN
IUCN_all <- read.csv(paste0(data.path, "Data/IUCN/Taxonomy_Sept2024.csv"))

all_sp <- all_sp %>% mutate(IUCN_name = case_when(scientific_name == "Aegithalos exilis" ~ "Psaltria exilis",
                                       scientific_name == "Aerodramus ceramensis" ~ "Aerodramus infuscatus",
                                       scientific_name == "Aerodramus germani" ~ "Aerodramus fuciphagus",
                                       scientific_name == "Aerodramus sororum" ~ "Aerodramus infuscatus",
                                       scientific_name == "Ailuroedus arfakianus" ~ "Ailuroedus melanotis",
                                       scientific_name == "Ailuroedus geislerorum" ~ "Ailuroedus buccoides",
                                       scientific_name == "Anarhynchus alexandrinus" ~ "Charadrius alexandrinus",
                                       scientific_name =="Anarhynchus atrifrons" ~ "Charadrius atrifrons",
                                       scientific_name == "Anarhynchus dealbatus" ~ "Charadrius dealbatus",
                                       scientific_name == "Anarhynchus javanicus" ~ "Charadrius javanicus",
                                       scientific_name == "Anarhynchus leschenaultii" ~ "Charadrius leschenaultii",
                                       scientific_name == "Anarhynchus mongolus" ~ "Charadrius mongolus",
                                       scientific_name == "Anarhynchus peronii" ~ "Charadrius peronii",
                                       scientific_name == "Anarhynchus ruficapillus" ~ "Charadrius ruficapillus",
                                       scientific_name == "Anarhynchus veredus" ~ "Charadrius veredus",
                                       scientific_name == "Anthus ruficollis" ~ "Madanga ruficollis",
                                       scientific_name == "Antigone rubicunda" ~ "Grus rubicunda",
                                       scientific_name == "Arachnothera everetti" ~ "Arachnothera affinis",
                                       scientific_name == "Artamus leucorynchus" ~ "Artamus leucoryn",
                                       scientific_name == "Batrachostomus affinis" ~ "Batrachostomus javensis",
                                       scientific_name == "Bubulcus coromandus" ~ "Bubulcus ibis",
                                       scientific_name == "Cacomantis leucolophus" ~ "Caliechthrus leucolophus",
                                       scientific_name == "Cacomantis pallidus" ~ "Heteroscenes pallidus",
                                       scientific_name == "Cecropis striolata" ~ "Cecropis daurica",
                                       scientific_name == "Celebesia abbotti" ~ "Celebesica abbotti",
                                       scientific_name == "Celebesia parvula" ~ "Coracina parvula",
                                       scientific_name == "Charadrius melanops" ~ "Elseyornis melanops",
                                       scientific_name == "Charminetta wilhelminae" ~ "Charmosyna wilhelminae",
                                       scientific_name == "Charmosynopsis pulchella" ~ "Charmosyna pulchella",
                                       scientific_name == "Chroicocephalus novaehollandiae" ~ "Larus novaehollandiae",
                                       scientific_name == "Chroicocephalus ridibundus" ~ "Larus ridibundus",
                                       scientific_name == "Chrysococcyx basalis" ~ "Chalcites basalis",
                                       scientific_name == "Chrysococcyx lucidus" ~ "Chalcites lucidus",
                                       scientific_name == "Chrysococcyx megarhynchus" ~ "Chalcites megarhynchus",
                                       scientific_name == "Chrysococcyx meyerii" ~ "Chalcites meyerii",
                                       scientific_name == "Chrysococcyx minutillus" ~ "Chalcites minutillus",
                                       scientific_name == "Chrysococcyx ruficollis" ~ "Chalcites ruficollis",
                                       scientific_name == "Cinnyris clementiae" ~ "Cinnyris jugularis",
                                       scientific_name == "Cinnyris frenatus" ~ "Cinnyris jugularis",
                                       scientific_name == "Cinnyris infrenatus" ~ "Cinnyris jugularis",
                                       scientific_name == "Cinnyris ornatus" ~ "Cinnyris jugularis",
                                       scientific_name == "Cinnyris teysmanni" ~ "Cinnyris jugularis",
                                       scientific_name == "Collocalia affinis" ~ "Collocalia esculenta",
                                       scientific_name == "Collocalia neglecta" ~ "Collocalia esculenta",
                                       scientific_name == "Collocalia sumbawae" ~ "Collocalia esculenta",
                                       scientific_name == "Colluricincla affinis" ~ "Colluricincla megarhyncha",
                                       scientific_name == "Colluricincla obscura" ~ "Colluricincla megarhyncha",
                                       scientific_name == "Copsychus stricklandii" ~ "Copsychus malabaricus",
                                       scientific_name == "Crateroscelis murina" ~ "Origma murina",
                                       scientific_name == "Crateroscelis robusta" ~ "Origma robusta",
                                       scientific_name == "Cyornis kalaoensis" ~ "Cyornis omissus",
                                       scientific_name == "Diphyllodes magnificus" ~ "Cicinnurus magnificus",
                                       scientific_name == "Diphyllodes respublica" ~ "Cicinnurus respublica",
                                       scientific_name == "Eopsaltria leucops" ~ "Tregellasia leucops",
                                       scientific_name == "Eopsaltria placens" ~ "Gennaeodryas placens",
                                       scientific_name == "Erythropitta habenichti" ~ "Erythropitta macklotii",
                                       scientific_name == "Eudynamys melanorhynchus" ~ "Eudynamys orientalis",
                                       scientific_name == "Gallirallus philippensis" ~ "Hypotaenidia philippensis",
                                       scientific_name == "Gallirallus torquatus" ~ "Hypotaenidia torquata",
                                       scientific_name == "Gallirallus wallacii" ~ "Habroptila wallacii",
                                       scientific_name == "Gecinulus rafflesii" ~ "Chloropicoides rafflesii",
                                       scientific_name == "Glossoptilus goldiei" ~ "Psitteuteles goldiei",
                                       scientific_name == "Gypsophila rufipectus" ~ "Gypsophila rufipecta",
                                       scientific_name == "Heteromyias brachyurus" ~ "Poecilodryas brachyura",
                                       scientific_name == "Himantopus leucocephalus" ~ "Himantopus himantopus",
                                       scientific_name == "Hypocharmosyna placentis" ~ "Charmosyna placentis",
                                       scientific_name == "Hypocharmosyna rubronotata" ~ "Charmosyna rubronotata",
                                       scientific_name == "Icthyophaga leucogaster" ~ "Haliaeetus leucogaster",
                                       scientific_name == "Iole crypta" ~ "Iole charlottae",
                                       scientific_name == "Ixos leucogrammicus" ~ "Hemixos leucogrammicus",
                                       scientific_name == "Ketupa sumatrana" ~ "Bubo sumatranus",
                                       scientific_name == "Kurochkinegramma hypogrammicum" ~ "Kurochkinegramma hypogrammica",
                                       scientific_name == "Leucoptilon concretum" ~ "Cyornis concretus",
                                       scientific_name == "Lophorina niedda" ~ "Lophorina superba",
                                       scientific_name == "Macropygia doreya" ~ "Macropygia amboinensis",
                                       scientific_name == "Macropygia modiglianii" ~ "Macropygia emiliana",
                                       scientific_name == "Mayrimunia leucosticta" ~ "Lonchura leucosticta",
                                       scientific_name == "Mayrimunia tristissima" ~ "Lonchura tristissima",
                                       scientific_name == "Melanocharis crassirostris" ~ "Rhamphocharis crassirostris",
                                       scientific_name == "Melanodryas cryptoleuca" ~ "Peneothello cryptoleuca",
                                       scientific_name == "Melanodryas cyanus" ~ "Peneothello cyanus",
                                       scientific_name == "Melanodryas pulverulenta" ~ "Peneoenanthe pulverulenta",
                                       scientific_name == "Melanodryas sigillata" ~ "Peneothello sigillata",
                                       scientific_name =="Microtarsus eutilotus" ~ "Euptilotus eutilotus",
                                       scientific_name =="Microtarsus melanocephalos" ~ "Brachypodius atriceps",
                                       scientific_name =="Muscicapa sodhii" ~ "Muscicapa dauurica",
                                       scientific_name == "Muscicapa williamsoni" ~ "Muscicapa dauurica",
                                       scientific_name == "Myzomela irianawidodoae" ~ "Myzomela dammermani",
                                       scientific_name == "Myzomela wahe" ~ "Myzomela chloroptera",
                                       scientific_name == "Pachycephala balim" ~ "Pachycephala pectoralis",
                                       scientific_name == "Pachycephala macrorhyncha" ~ "Pachycephala pectoralis",
                                       scientific_name == "Pachycephala melanorhyncha" ~ "Colluricincla megarhyncha",
                                       scientific_name == "Parus cinereus" ~ "Parus major",
                                       scientific_name == "Pellorneum capistratoides" ~ "Pellorneum nigrocapitatum",
                                       scientific_name == "Pericrocotus speciosus" ~ "Pericrocotus flammeus",
                                       scientific_name == "Pitta forsteni" ~ "Pitta sordida",
                                       scientific_name == "Poliolimnas cinereus" ~ "Amaurornis cinerea",
                                       scientific_name == "Porphyrio indicus" ~ "Porphyrio porphyrio",
                                       scientific_name == "Porphyrio melanotus" ~ "Porphyrio porphyrio",
                                       scientific_name =="Porphyrio poliocephalus" ~ "Muscicapa dauurica",
                                       scientific_name == "Psittacula longicauda" ~ "Belocercus longicaudus",
                                       scientific_name == "Ptilinopus bernsteinii" ~ "Megaloprepia formosa",
                                       scientific_name == "Ptilinopus epius" ~ "Ramphiculus epius",
                                       scientific_name == "Ptilinopus fischeri" ~ "Ramphiculus fischeri",
                                       scientific_name == "Ptilinopus jambu" ~ "Ramphiculus jambu",
                                       scientific_name == "Ptilinopus magnificus" ~ "Megaloprepia magnifica",
                                       scientific_name == "Ptilinopus mangoliensis" ~ "Ramphiculus mangoliensis",
                                       scientific_name == "Ptilinopus subgularis" ~ "Ramphiculus subgularis",
                                       scientific_name == "Ptiloris magnificus" ~ "Lophorina magnifica",
                                       scientific_name == "Rhipidura habibiei" ~ "Rhipidura teysmanni", #
                                       scientific_name == "Rhipidura semicollaris" ~ "Rhipidura dryas",
                                       scientific_name == "Rhipidura sulaensis" ~ "Rhipidura teysmanni",
                                       scientific_name == "Rhipidura torrida" ~ "Rhipidura rufifrons",
                                       scientific_name == "Rubigula cyaniventris" ~ "Ixidia cyaniventris",
                                       scientific_name == "Rubigula erythropthalmos" ~ "Ixidia erythropthalmos",
                                       scientific_name == "Rubigula squamata" ~ "Ixidia squamata",
                                       scientific_name == "Saudareos flavoviridis" ~ "Trichoglossus flavoviridis",
                                       scientific_name == "Saudareos iris" ~ "Trichoglossus iris",
                                       scientific_name == "Saudareos meyeri" ~ "Trichoglossus meyeri",
                                       scientific_name == "Saudareos ornata" ~ "Trichoglossus ornatus",
                                       scientific_name == "Synoicus monorthonyx" ~ "Anurophasis monorthonyx",
                                       scientific_name == "Taenioptynx sylvaticus" ~ "Glaucidium sylvaticum",
                                       scientific_name == "Territornis reticulata" ~ "Microptilotis reticulatus",
                                       scientific_name == "Threskiornis molucca" ~ "Threskiornis moluccus",
                                       scientific_name == "Todiramphus sordidus" ~ "Todiramphus chloris",
                                       scientific_name == "Tyto almae" ~ "Tyto novaehollandiae",
                                       scientific_name == "Tyto sororcula" ~ "Tyto novaehollandiae",
                                       scientific_name == "Yungipicus canicapillus" ~ "Picoides canicapillus",
                                       scientific_name == "Yungipicus moluccensis" ~ "Picoides moluccensis",
                                       scientific_name == "Yungipicus temminckii" ~ "Picoides temminckii",
                                       # added as subsequently found this species was split in the IUCN
                                       # and tahitica represents just the tahiti pop per the IUCN whereas
                                       # eBird use that for the wider pacifci pop
                                       scientific_name == "Hirundo tahitica" ~ "Hirundo javanica",
                                       ## MY only sp.
                                       scientific_name == "Actinodura cyanouroptera" ~ "Siva cyanouroptera",
                                       scientific_name == "Actinodura strigula" ~ "Chrysominla strigula",
                                       scientific_name == "Apus cooki" ~ "Apus pacificus",
                                       scientific_name == "Cecropis badia" ~ "Cecropis daurica",
                                       scientific_name == "Chroicocephalus brunnicephalus" ~ "Larus brunnicephalus",
                                       scientific_name == "Chroicocephalus genei" ~ "Larus genei",
                                       scientific_name == "Collocalia dodgei" ~ "Collocalia linchi",
                                       scientific_name == "Enicurus borneensis" ~ "Enicurus leschenaulti",
                                       scientific_name == "Ketupa coromanda" ~ "Bubo coromandus",
                                       scientific_name == "Psittacula krameri" ~ "Alexandrinus krameri",
                                       scientific_name == "Pycnonotus conradi" ~ "Pycnonotus blanfordi",
                                       scientific_name == "Saxicola stejnegeri" ~ "Saxicola torquatus",
                                       scientific_name == "Taenioptynx brodiei" ~ "Glaucidium brodiei",
                                       .default = scientific_name))

IUCN_sub <- IUCN_all %>% select(scientificName, redlistCategory, populationTrend)

match <- left_join(all_sp, IUCN_sub, by = c("IUCN_name" = "scientificName"))
match %>% filter(is.na(redlistCategory))

write.csv(match, paste0(data.path, "Data/Taxonomy/eBird_IUCN_ID.MY_match.csv"))

# AVONET taxonomy

avonet_bl_raw <- read.csv(paste0(data.path, 
                                 "Data/Traits/AVONET/AVONET_Birdlife_Taxonomoy_RAW.csv"))

avonet_taxo2 <- avonet_bl_raw %>% select(Species1, Mass)


all_avo_sub <- match %>% 
  mutate(AVONET_BL_name = case_when(IUCN_name == "Aethomyias arfakianus" ~ "Sericornis arfakianus",
                                 IUCN_name == "Aethomyias nigrorufus" ~ "Crateroscelis nigrorufa",
                                 IUCN_name == "Aethomyias papuensis" ~ "Sericornis papuensis",
                                 IUCN_name == "Aethomyias perspicillatus" ~ "Sericornis perspicillatus",
                                 IUCN_name == "Aethomyias rufescens" ~ "Sericornis rufescens",
                                 IUCN_name == "Aethomyias spilodera" ~ "Sericornis spilodera",
                                 IUCN_name == "Charadrius atrifrons" ~ "Charadrius mongolus",
                                 IUCN_name == "Grus rubicunda" ~ "Antigone rubicunda",
                                 IUCN_name == "Cacatua citrinocristata" ~ "Cacatua sulphurea",
                                 IUCN_name == "Coracina parvula" ~ "Edolisoma parvulum",
                                 IUCN_name == "Ceyx rufidorsa" ~ "Ceyx erithaca",
                                 IUCN_name == "Cincloramphus bivittatus" ~ "Buettikoferella bivittata",
                                 IUCN_name == "Copsychus malabaricus" ~ "Kittacincla malabarica",
                                 IUCN_name == "Copsychus pyrropygus" ~ "Trichixos pyrropygus",
                                 IUCN_name == "Origma murina" ~ "Crateroscelis murina",
                                 IUCN_name == "Origma robusta" ~ "Crateroscelis robusta",
                                 IUCN_name == "Cuculus optatus" ~ "Cuculus saturatus",
                                 IUCN_name == "Cyornis kadayangensis" ~ "Cyornis banyumas", #new sp, closest to, per IUCN
                                 IUCN_name == "Cyornis pelingensis" ~ "Cyornis colonus",
                                 IUCN_name == "Dacelo rex" ~ "Clytoceyx rex",
                                 IUCN_name == "Eumyias hoevelli" ~ "Cyornis hoevelli",
                                 IUCN_name == "Eumyias hyacinthinus" ~ "Cyornis hyacinthinus",
                                 IUCN_name == "Eumyias oscillans" ~ "Cyornis oscillans",
                                 IUCN_name == "Eumyias sanfordi" ~ "Cyornis sanfordi",
                                 IUCN_name == "Eumyias stresemanni" ~ "Cyornis stresemanni",
                                 IUCN_name == "Chloropicoides rafflesii" ~ "Dinopium rafflesii",
                                 IUCN_name == "Gypsophila rufipecta" ~ "Turdinus rufipectus",
                                 IUCN_name == "Helopsaltes amnicola" ~ "Locustella amnicola",
                                 IUCN_name == "Helopsaltes certhiola" ~ "Locustella certhiola",
                                 IUCN_name == "Helopsaltes fasciolatus" ~ "Locustella fasciolata",
                                 IUCN_name == "Locustella disturbans" ~ "Locustella castanea",
                                 IUCN_name == "Locustella musculus" ~ "Locustella castanea",
                                 IUCN_name == "Macropygia cinnamomea" ~ "Macropygia emiliana",
                                 IUCN_name == "Melanocichla calva" ~ "Garrulax calvus",
                                 IUCN_name == "Melanocichla lugubris" ~ "Garrulax lugubris",
                                 IUCN_name == "Melionyx fuscus" ~ "Melidectes fuscus",
                                 IUCN_name == "Melionyx nouhuysi" ~ "Melidectes nouhuysi",
                                 IUCN_name == "Napothera albostriata" ~ "Rimator albostriatus",
                                 IUCN_name == "Ninox fusca" ~ "Ninox boobook",
                                 IUCN_name == "Ninox plesseni" ~ "Ninox boobook",
                                 IUCN_name == "Ninox rotiensis" ~ "Ninox boobook",
                                 IUCN_name == "Padda fuscata" ~ "Lonchura fuscata",
                                 IUCN_name == "Padda oryzivora" ~ "Lonchura oryzivora",
                                 IUCN_name == "Pampusana beccarii" ~ "Alopecoenas beccarii",
                                 IUCN_name == "Pampusana hoedtii" ~ "Alopecoenas hoedtii",
                                 IUCN_name == "Pampusana jobiensis" ~ "Alopecoenas jobiensis",
                                 IUCN_name == "Pellorneum bicolor" ~ "Trichastoma bicolor",
                                 IUCN_name == "Pellorneum buettikoferi" ~ "Trichastoma buettikoferi",
                                 IUCN_name == "Pellorneum celebense" ~ "Trichastoma celebense",
                                 IUCN_name == "Pellorneum malaccense" ~ "Trichastoma malaccense",
                                 IUCN_name == "Pellorneum pyrrogenys" ~ "Trichastoma pyrrogenys",
                                 IUCN_name == "Pellorneum rostratum" ~ "Trichastoma rostratum",
                                 IUCN_name == "Phylloscopus nesophilus" ~ "Phylloscopus sarasinorum",
                                 IUCN_name == "Pitta concinna" ~ "Pitta elegans",
                                 IUCN_name == "Pitta vigorsii" ~ "Pitta elegans",
                                 IUCN_name == "Pomatorhinus bornensis" ~ "Pomatorhinus montanus",
                                 IUCN_name == "Pterorhinus mitratus" ~ "Garrulax mitratus",
                                 IUCN_name == "Pterorhinus treacheri" ~ "Garrulax treacheri",
                                 IUCN_name == "Staphida everetti" ~ "Yuhina everetti",
                                 IUCN_name == "Glaucidium sylvaticum" ~ "Glaucidium brodiei",
                                 IUCN_name == "Zoothera heinrichi" ~ "Geomalia heinrichi",
                                 IUCN_name == "Zosterops dehaani" ~ "Zosterops atriceps",
                                 IUCN_name == "Zosterops flavissimus" ~ "Zosterops chloris",
                                 IUCN_name == "Zosterops meratusensis" ~ "Zosterops chloris", # new sp. closest co
                                 IUCN_name == "Zosterops paruhbesar" ~ "Zosterops chloris", # new sp. geographically only co on island
                                 ## MY only sp.
                                 IUCN_name == "Cyornis montanus" ~ "Cyornis banyumas",
                                 IUCN_name == "Cyornis whitei" ~ "Cyornis banyumas",
                                 IUCN_name == "Gracupica floweri" ~ "Gracupica contra",
                                 IUCN_name == "Gypsophila brevicaudata" ~ "Turdinus brevicaudatus",
                                 IUCN_name == "Gypsophila crassa" ~ "Turdinus crassus",
                                 IUCN_name == "Helopsaltes ochotensis" ~ "Locustella ochotensis",
                                 IUCN_name == "Luscinia svecica" ~ "Cyanecula svecica",
                                 IUCN_name == "Oriolus consobrinus" ~ "Oriolus xanthonotus",
                                 IUCN_name == "Pellorneum tickelli" ~ "Trichastoma tickelli",
                                 IUCN_name == "Tropicoperdix charltonii" ~ "Arborophila charltonii",
                                 IUCN_name == "Tropicoperdix graydoni" ~ "Arborophila graydoni",
                                  .default = IUCN_name))

test <- all_avo_sub %>% left_join(avonet_taxo2, by = c("AVONET_BL_name" = "Species1"))
test %>% filter(is.na(Mass)) %>% select(scientific_name, common_name, IUCN_name)

write.csv(all_avo_sub, paste0(data.path, 
                              "Data/Taxonomy/eBird_IUCN_AVONET_ID.MY_match.csv"))
